# docs/specs/swagger.yaml
{"openapi":"3.0.1","info":{"title":"finalProject","description":"bootcamp-br-gp9-project","contact":{"name":"finalProject","email":"ext_dambetan@mercadolibre.com"},"version":"last"},"servers":[{"url":"https://null_finalProject.furyapps.io","description":"Scope null"}],"paths":{"/seller":{"get":{"tags":["seller-controller"],"operationId":"listAllSellers","responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/ApiError"},{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}]}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Seller"}}}}}}},"put":{"tags":["seller-controller"],"operationId":"updateSeller","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Seller"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/ApiError"},{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}]}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Seller"}}}}}},"post":{"tags":["seller-controller"],"operationId":"createSeller","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Seller"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/ApiError"},{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}]}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"integer","format":"int64"}}}}}}},"/api/v1/fresh-products/inboundorder":{"put":{"tags":["inbound-order-controller"],"operationId":"update","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InboundOrderUpdateRequestDTO"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/ApiError"},{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}]}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/InboundOrderResponseDTO"}}}}}},"post":{"tags":["inbound-order-controller"],"operationId":"create_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InboundOrderCreateRequestDTO"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/ApiError"},{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}]}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/InboundOrderResponseDTO"}}}}}}},"/api/v1/sign-in":{"post":{"tags":["session-controller"],"operationId":"login","parameters":[{"name":"username","in":"query","required":true,"schema":{"type":"string"}},{"name":"password","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/ApiError"},{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}]}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AccountResponseDTO"}}}}}}},"/api/v1/fresh-products/purchase-order":{"get":{"tags":["purchase-order-controller"],"operationId":"getAll","responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/ApiError"},{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}]}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PurchaseOrderResponseDTO"}}}}}}},"post":{"tags":["purchase-order-controller"],"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PurchaseOrderRequestDTO"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/ApiError"},{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}]}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PurchaseOrderResponseDTO"}}}}}}},"/api/v1/fresh-products/purchase-order/update":{"post":{"tags":["purchase-order-controller"],"operationId":"update_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PurchaseOrderUpdateRequestDTO"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/ApiError"},{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}]}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PurchaseOrderResponseDTO"}}}}}}},"/api/v1/countries":{"post":{"tags":["country-controller"],"operationId":"create_2","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CountryRequestDTO"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/ApiError"},{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}]}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CountryResponseDTO"}}}}}}},"/seller/{id}":{"get":{"tags":["seller-controller"],"operationId":"findSellerById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/ApiError"},{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}]}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Seller"}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"operationId":"ping","responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/ApiError"},{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}]}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/fresh-products":{"get":{"tags":["product-controller"],"operationId":"getProductsByCountry","parameters":[{"name":"productType","in":"query","required":false,"schema":{"type":"integer","format":"int32"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/ApiError"},{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}]}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductResponseDTO"}}}}}}}},"/api/v1/fresh-products/warehouse":{"get":{"tags":["product-sum-controller"],"operationId":"getSumOfProductInAllWarehouses","parameters":[{"name":"productId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"countryId","in":"query","required":false,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/ApiError"},{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}]}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SumOfProductStockDTO"}}}}}}},"/api/v1/fresh-products/purchase-order/{id}":{"get":{"tags":["purchase-order-controller"],"operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/ApiError"},{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}]}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PurchaseOrderResponseDTO"}}}}}}},"/api/v1/fresh-products/list":{"get":{"tags":["product-controller"],"operationId":"getSectorBatchesByProductId","parameters":[{"name":"productId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"ordered","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/ApiError"},{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}]}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SectorBatchResponseDTO"}}}}}}},"/api/v1/fresh-products/due-date":{"get":{"tags":["batch-controller"],"operationId":"getBatchesBySector","parameters":[{"name":"sectorId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"daysQuantity","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/ApiError"},{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}]}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BatchValidateDateResponseDTO"}}}}}}}},"/api/v1/fresh-products/due-date/list":{"get":{"tags":["batch-controller"],"operationId":"getBatchesByProductType","parameters":[{"name":"daysQuantity","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"category","in":"query","required":true,"schema":{"type":"string"}},{"name":"direction","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/ApiError"},{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}]}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BatchValidateDateResponseDTO"}}}}}}}}},"components":{"schemas":{"ValidationError":{"type":"object","properties":{"field":{"type":"string"},"message":{"type":"string"}}},"ApiError":{"type":"object","properties":{"error":{"type":"string"},"message":{"type":"string"},"status":{"type":"integer","format":"int32"},"errors":{"type":"array","items":{"$ref":"#/components/schemas/SubError"}}}},"SubError":{"type":"object","properties":{"cause":{"type":"object","properties":{"stackTrace":{"type":"array","items":{"type":"object","properties":{"classLoaderName":{"type":"string"},"moduleName":{"type":"string"},"moduleVersion":{"type":"string"},"methodName":{"type":"string"},"fileName":{"type":"string"},"lineNumber":{"type":"integer","format":"int32"},"className":{"type":"string"},"nativeMethod":{"type":"boolean"}}}},"message":{"type":"string"},"suppressed":{"type":"array","items":{"type":"object","properties":{"stackTrace":{"type":"array","items":{"type":"object","properties":{"classLoaderName":{"type":"string"},"moduleName":{"type":"string"},"moduleVersion":{"type":"string"},"methodName":{"type":"string"},"fileName":{"type":"string"},"lineNumber":{"type":"integer","format":"int32"},"className":{"type":"string"},"nativeMethod":{"type":"boolean"}}}},"message":{"type":"string"},"localizedMessage":{"type":"string"}}}},"localizedMessage":{"type":"string"}}},"stackTrace":{"type":"array","items":{"type":"object","properties":{"classLoaderName":{"type":"string"},"moduleName":{"type":"string"},"moduleVersion":{"type":"string"},"methodName":{"type":"string"},"fileName":{"type":"string"},"lineNumber":{"type":"integer","format":"int32"},"className":{"type":"string"},"nativeMethod":{"type":"boolean"}}}},"message":{"type":"string"},"suppressed":{"type":"array","items":{"type":"object","properties":{"stackTrace":{"type":"array","items":{"type":"object","properties":{"classLoaderName":{"type":"string"},"moduleName":{"type":"string"},"moduleVersion":{"type":"string"},"methodName":{"type":"string"},"fileName":{"type":"string"},"lineNumber":{"type":"integer","format":"int32"},"className":{"type":"string"},"nativeMethod":{"type":"boolean"}}}},"message":{"type":"string"},"localizedMessage":{"type":"string"}}}},"localizedMessage":{"type":"string"}}},"Account":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"username":{"type":"string"},"password":{"type":"string"},"rol":{"type":"integer","format":"int32"},"country":{"$ref":"#/components/schemas/Country"}}},"Batch":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"product":{"$ref":"#/components/schemas/Product"},"sector":{"$ref":"#/components/schemas/Sector"},"inboundOrder":{"$ref":"#/components/schemas/InboundOrder"},"currentTemperature":{"type":"number","format":"float"},"minimumTemperature":{"type":"number","format":"float"},"initialQuantity":{"type":"integer","format":"int32"},"currentQuantity":{"type":"integer","format":"int32"},"manufacturingDate":{"type":"string","format":"date"},"manufacturingTime":{"type":"string","format":"date-time"},"dueDate":{"type":"string","format":"date"},"sectortype":{"type":"integer","format":"int32"}}},"Country":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"accounts":{"type":"array","items":{"$ref":"#/components/schemas/Account"}},"warehouses":{"type":"array","items":{"$ref":"#/components/schemas/Warehouse"}}}},"InboundOrder":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"orderDate":{"type":"string","format":"date"},"batches":{"type":"array","items":{"$ref":"#/components/schemas/Batch"}},"representative":{"$ref":"#/components/schemas/Representative"}}},"Product":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"description":{"type":"string"},"price":{"type":"number","format":"double"},"productType":{"type":"integer","format":"int32"},"seller":{"$ref":"#/components/schemas/Seller"}}},"Representative":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"warehouse":{"$ref":"#/components/schemas/Warehouse"},"account":{"$ref":"#/components/schemas/Account"}}},"Sector":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"sectorType":{"type":"integer","format":"int32"},"batches":{"type":"array","items":{"$ref":"#/components/schemas/Batch"}},"warehouse":{"$ref":"#/components/schemas/Warehouse"},"maxQuantityBatches":{"type":"integer","format":"int32"},"wareHouseId":{"type":"integer","format":"int64"}}},"Seller":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"products":{"type":"array","items":{"$ref":"#/components/schemas/Product"}},"account":{"$ref":"#/components/schemas/Account"}}},"Warehouse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"country":{"$ref":"#/components/schemas/Country"},"sectors":{"type":"array","items":{"$ref":"#/components/schemas/Sector"}},"representative":{"$ref":"#/components/schemas/Representative"}}},"BatchRequestUpdateDTO":{"required":["batchNumber","currentQuantity","currentTemperature","dueDate","initialQuantity","manufacturingDate","manufacturingTime","minimumTemperature","productId"],"type":"object","properties":{"productId":{"type":"integer","format":"int64"},"currentTemperature":{"type":"number","format":"float"},"minimumTemperature":{"type":"number","format":"float"},"initialQuantity":{"type":"integer","format":"int32"},"currentQuantity":{"type":"integer","format":"int32"},"manufacturingDate":{"type":"string","format":"date"},"manufacturingTime":{"type":"string","format":"date-time"},"dueDate":{"type":"string","format":"date"},"batchNumber":{"type":"integer","format":"int64"}}},"InboundOrderUpdateRequestDTO":{"required":["batchStock","orderDate","orderNumber","section"],"type":"object","properties":{"orderNumber":{"type":"integer","format":"int64"},"orderDate":{"type":"string","format":"date"},"section":{"$ref":"#/components/schemas/SectorRequestDTO"},"batchStock":{"type":"array","items":{"$ref":"#/components/schemas/BatchRequestUpdateDTO"}},"username":{"type":"string"}}},"SectorRequestDTO":{"required":["sectionCode","warehouseCode"],"type":"object","properties":{"warehouseCode":{"type":"integer","format":"int64"},"sectionCode":{"type":"integer","format":"int64"}}},"BatchDTO":{"type":"object","properties":{"productId":{"type":"integer","format":"int64"},"currentTemperature":{"type":"number","format":"float"},"minimumTemperature":{"type":"number","format":"float"},"initialQuantity":{"type":"integer","format":"int32"},"currentQuantity":{"type":"integer","format":"int32"},"manufacturingDate":{"type":"string","format":"date"},"manufacturingTime":{"type":"string","format":"date-time"},"dueDate":{"type":"string","format":"date"},"batchNumber":{"type":"integer","format":"int64"}}},"InboundOrderResponseDTO":{"type":"object","properties":{"orderId":{"type":"integer","format":"int64"},"batchStock":{"type":"array","items":{"$ref":"#/components/schemas/BatchDTO"}}}},"AccountResponseDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"username":{"type":"string"},"password":{"type":"string"},"token":{"type":"string"},"country":{"$ref":"#/components/schemas/CountryResponseDTO"}}},"CountryResponseDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"}}},"ProductPurchaseOrderRequestDTO":{"type":"object","properties":{"productId":{"type":"integer","format":"int64"},"quantity":{"type":"integer","format":"int32"}}},"PurchaseOrderRequestDTO":{"type":"object","properties":{"date":{"type":"string","format":"date"},"buyerId":{"type":"integer","format":"int64"},"orderStatus":{"type":"integer","format":"int32"},"products":{"type":"array","items":{"$ref":"#/components/schemas/ProductPurchaseOrderRequestDTO"}}}},"BatchPurchaseOrderResponseDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"batchId":{"type":"integer","format":"int64"},"quantity":{"type":"integer","format":"int32"},"manufacturingDate":{"type":"string","format":"date"},"manufacturingTime":{"type":"string","format":"date-time"},"dueDate":{"type":"string","format":"date"}}},"ProductBatchesPurchaseOrderResponseDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"productId":{"type":"integer","format":"int64"},"productName":{"type":"string"},"quantity":{"type":"integer","format":"int32"},"price":{"type":"number","format":"double"},"batches":{"type":"array","items":{"$ref":"#/components/schemas/BatchPurchaseOrderResponseDTO"}}}},"PurchaseOrderResponseDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"date":{"type":"string","format":"date"},"totalPrice":{"type":"number","format":"double"},"products":{"type":"array","items":{"$ref":"#/components/schemas/ProductBatchesPurchaseOrderResponseDTO"}}}},"PurchaseOrderUpdateRequestDTO":{"type":"object","properties":{"purchaseOrderId":{"type":"integer","format":"int64"},"productId":{"type":"integer","format":"int64"},"newQuantity":{"type":"integer","format":"int32"}}},"BatchRequestCreateDTO":{"required":["currentQuantity","currentTemperature","dueDate","initialQuantity","manufacturingDate","manufacturingTime","minimumTemperature","productId"],"type":"object","properties":{"productId":{"type":"integer","format":"int64"},"currentTemperature":{"type":"number","format":"float"},"minimumTemperature":{"type":"number","format":"float"},"initialQuantity":{"type":"integer","format":"int32"},"currentQuantity":{"type":"integer","format":"int32"},"manufacturingDate":{"type":"string","format":"date"},"manufacturingTime":{"type":"string","format":"date-time"},"dueDate":{"type":"string","format":"date"}}},"InboundOrderCreateRequestDTO":{"required":["batchStock","orderDate","section"],"type":"object","properties":{"orderDate":{"type":"string","format":"date"},"section":{"$ref":"#/components/schemas/SectorRequestDTO"},"batchStock":{"type":"array","items":{"$ref":"#/components/schemas/BatchRequestCreateDTO"}},"username":{"type":"string"}}},"CountryRequestDTO":{"required":["name"],"type":"object","properties":{"name":{"maxLength":2147483647,"minLength":3,"type":"string"}}},"ProductResponseDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"description":{"type":"string"},"price":{"type":"number","format":"double"},"type":{"type":"integer","format":"int32"},"typeDescription":{"type":"string"}}},"SumOfProductStockDTO":{"type":"object","properties":{"productId":{"type":"integer","format":"int64"},"warehouses":{"type":"array","items":{"$ref":"#/components/schemas/WarehouseProductSumDTO"}}}},"WarehouseProductSumDTO":{"type":"object","properties":{"warehouseCode":{"type":"integer","format":"int64"},"totalQuantity":{"type":"integer","format":"int32"}}},"BatchIdentificationResponseDTO":{"type":"object","properties":{"sectorCode":{"type":"integer","format":"int64"},"warehouseCode":{"type":"integer","format":"int64"}}},"BatchStockResponseDTO":{"type":"object","properties":{"batchNumber":{"type":"integer","format":"int64"},"currentQuantity":{"type":"integer","format":"int32"},"dueDate":{"type":"string","format":"date"}}},"SectorBatchResponseDTO":{"type":"object","properties":{"sector":{"$ref":"#/components/schemas/BatchIdentificationResponseDTO"},"productId":{"type":"integer","format":"int64"},"batchStock":{"type":"array","items":{"$ref":"#/components/schemas/BatchStockResponseDTO"}}}},"BatchValidateDateResponseDTO":{"type":"object","properties":{"batchNumber":{"type":"integer","format":"int64"},"productId":{"type":"integer","format":"int64"},"productTypeId":{"type":"integer","format":"int32"},"dueDate":{"type":"string","format":"date"},"quantity":{"type":"integer","format":"int32"}}}}}}